refactor(db): move db folder in place
