+++ drizzle.config.ts
@@ -1,6 +1,6 @@
 const config = {
   dialect: "postgresql",
-  schema: "./src/db/schema.ts",
+  schema: "./src/server/db/schema.ts",
   dbCredentials: {
     url: process.env.DATABASE_URL,
   },
