+++ package.json
@@ -8,8 +8,8 @@
     "start": "next start",
     "lint": "next lint",
     "generate": "drizzle-kit generate",
-    "migrate:up": "node ./src/db/migrate.js",
-    "seed": "node --loader esbuild-register/loader -r esbuild-register ./src/db/seed/index.ts"
+    "migrate:up": "node ./src/server/db/migrate.js",
+    "seed": "node --loader esbuild-register/loader -r esbuild-register ./src/server/db/seed/index.ts"
   },
   "dependencies": {
     "drizzle-orm": "^0.32.1",
