+++ src/app/api/advocates/route.ts
@@ -1,6 +1,5 @@
-import db from "../../../db";
-import { advocates } from "../../../db/schema";
-import { advocateData } from "../../../db/seed/advocates";
+import db from "@/server/db";
+import { advocates } from "@/server/db/schema";
 
 export async function GET() {
   const data = await db.select().from(advocates);
