+++ src/app/api/seed/route.ts
@@ -1,6 +1,6 @@
-import db from "../../../db";
-import { advocates } from "../../../db/schema";
-import { advocateData } from "../../../db/seed/advocates";
+import db from "@/server/db";
+import { advocates } from "@/server/db/schema";
+import { advocateData } from "@/server/db/seed/advocates";
 
 export async function POST() {
   const records = await db.insert(advocates).values(advocateData).returning();
